Move _generate_uids to server.py to avoid cyclical imports
